backend/ai_service/requirements.txt
fastapi
uvicorn
pypdf
python-docx
google-generativeai
pytest
httpx
python-multipart
python-dotenv
